// deno_dist/interfaces.ts
export type BluetoothServiceUUID = number | string;
/** Characteristic UUID. */
export type BluetoothCharacteristicUUID = number | string;
/** Descriptor UUID. */
export type BluetoothDescriptorUUID = number | string;

/** Manufacturer-specific data. */
export type BluetoothManufacturerData = Map<number, DataView>;
/** Service-specific data. */
export type BluetoothServiceData = Map<BluetoothServiceUUID, DataView>;
// ...
export interface BluetoothDataFilter {
    readonly dataPrefix?: ArrayBuffer | undefined;
    readonly mask?: ArrayBuffer | undefined;
}

/** Filter by manufacturer data. */
export interface BluetoothManufacturerDataFilter extends BluetoothDataFilter {
    companyIdentifier: number;
}

/** Filter by service data. */
export interface BluetoothServiceDataFilter extends BluetoothDataFilter {
    service: BluetoothServiceUUID;
}

/** Filter for scanning Bluetooth devices. */
export interface BluetoothLEScanFilter {
    readonly name?: string | undefined;
    readonly namePrefix?: string | undefined;
    readonly services?: BluetoothServiceUUID[] | undefined;
    readonly manufacturerData?: BluetoothManufacturerDataFilter[] | undefined;
    readonly serviceData?: BluetoothServiceDataFilter[] | undefined;
}

/** Information for filtering devices during scanning. */
export interface RequestDeviceInfo {
    address: string;
    name?: string | undefined;
    manufacturerData?: BluetoothManufacturerData | undefined;
    services?: BluetoothServiceUUID[] | undefined;
}
// ...
export type RequestDeviceOptions = {
    filters: BluetoothLEScanFilter[];
    optionalServices?: BluetoothServiceUUID[] | undefined;
    optionalManufacturerData?: number[] | undefined;
    timeout?: number;
    signal?: AbortSignal;
} | {
    filter: (info: RequestDeviceInfo) => boolean;
    timeout?: number;
    signal?: AbortSignal;
} | {
    acceptAllDevices: boolean;
    optionalServices?: BluetoothServiceUUID[] | undefined;
    optionalManufacturerData?: number[] | undefined;
    timeout?: number;
    signal?: AbortSignal;
};